requirements.txt
Flask
gunicorn
celery
redis
requests
networkx
matplotlib
python-nmap